+++ database/migrations/2025_03_27_021343_create_web_settings_table.php
@@ -11,30 +11,33 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('web_settings', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->string('logo');
-            $table->string('email');
-            $table->string('phone');
-            $table->string('address');
-            $table->string('facebook');
-            $table->string('instagram');
-            $table->string('twitter');
-            $table->string('youtube');
-            $table->string('website');
-            $table->string('coursellink1');
-            $table->string('coursellink2');
-            $table->string('coursellink3');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('web_settings')) {
+
+            Schema::create('web_settings', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->string('logo');
+                $table->string('email');
+                $table->string('phone');
+                $table->string('address');
+                $table->string('facebook');
+                $table->string('instagram');
+                $table->string('twitter');
+                $table->string('youtube');
+                $table->string('website');
+                $table->string('coursellink1');
+                $table->string('coursellink2');
+                $table->string('coursellink3');
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('web_settings');
+        if (Schema::hasTable('web_settings')) {
+            Schema::dropIfExists('web_settings');
+        }
     }
+
 };
